+++ hassio-1password-addon/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- Fixed a silent error in Docker build
+
 ## [0.2.0] - 2026-02-10
 
 - Initial public release